rebase

+++ catalogs/catalog-jdbc-doris/src/test/java/com/datastrato/gravitino/catalog/doris/integration/test/CatalogDorisIT.java
@@ -123,7 +123,10 @@ public class CatalogDorisIT extends AbstractIT {
     NameIdentifier[] nameIdentifiers =
         catalog.asTableCatalog().listTables(Namespace.of(metalakeName, catalogName, schemaName));
     for (NameIdentifier nameIdentifier : nameIdentifiers) {
-      catalog.asTableCatalog().dropTable(nameIdentifier);
+      // The table [test_add_index_*]'s state is SCHEMA_CHANGE, cannot be dropped.
+      if (!nameIdentifier.name().contains("test_add_index")) {
+        catalog.asTableCatalog().dropTable(nameIdentifier);
+      }
     }
     catalog.asSchemas().dropSchema(schemaName, true);
   }
@@ -263,7 +266,10 @@ public class CatalogDorisIT extends AbstractIT {
             null);
 
     // Try to drop a database, and cascade equals to false, it should not be allowed.
-    Assertions.assertFalse(catalog.asSchemas().dropSchema(schemaName, false));
+    Throwable excep =
+        Assertions.assertThrows(
+            RuntimeException.class, () -> catalog.asSchemas().dropSchema(schemaName, false));
+    Assertions.assertTrue(excep.getMessage().contains("the value of cascade should be true."));
 
     // Check the database still exists
     catalog.asSchemas().loadSchema(schemaName);
